bug fixes 456

+++ app/src/main/java/ru/filit/motiv/app/models/main/UnitValue.kt
@@ -1,3 +1,3 @@
 package ru.filit.motiv.app.models.main
 
-data class UnitValue (var value: Int?, var unit: String?)
+data class UnitValue (var value: Double?, var unit: String?)
